+++ src/main/java/sg/edu/smu/cs301/group3/campaignms/beans/SpendBean.java
@@ -20,7 +20,7 @@ public class SpendBean {
 
     private String currency;
 
-    private Long amount;
+    private double amount;
 
     private Date transaction_date;
